requirements.txt
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.5
django-toolbelt==0.0.1
funcsigs==0.4
gunicorn==19.3.0
matplotlib==1.4.3
mock==1.3.0
nose==1.3.7
numexpr==2.4.4
numpy==1.10.1
pandas==0.17.0
patsy==0.4.0
pbr==1.8.1
psycopg2==2.6.1
pyparsing==2.0.3
python-dateutil==2.4.2
pytz==2015.6
scipy==0.16.0
six==1.10.0
static3==0.6.1
statsmodels==0.6.1
tables==3.2.2
wheel==0.24.0
xlrd==0.9.4
